+++ src/main/java/com/github/ars_zero/common/item/SpellcastingCirclet.java
@@ -1,5 +1,7 @@
 package com.github.ars_zero.common.item;
 
+import com.github.ars_zero.client.gui.AbstractMultiPhaseCastDeviceScreen;
+import com.github.ars_zero.client.gui.SpellcastingCircletGUI;
 import com.github.ars_zero.client.renderer.item.ArchmageSpellStaffRenderer;
 import com.hollingsworth.arsnouveau.client.registry.ModKeyBindings;
 import com.github.ars_zero.common.spell.MultiPhaseCastContext;
@@ -27,6 +29,12 @@ import java.util.function.Consumer;
 public class SpellcastingCirclet extends AbstractMultiPhaseCastDevice implements ICurioItem {
     public SpellcastingCirclet() {
         super(SpellTier.THREE, new Properties());
+    }
+    
+    @OnlyIn(Dist.CLIENT)
+    @Override
+    protected AbstractMultiPhaseCastDeviceScreen createDeviceScreen(ItemStack stack, InteractionHand hand) {
+        return new SpellcastingCircletGUI(stack, hand);
     }
     
     public void beginCurioCast(Player player, ItemStack stack) {
